+++ mmdet/models/roi_heads/scnet_roi_head.py
@@ -342,7 +342,7 @@ class SCNetRoIHead(CascadeRoIHead):
         if rois.shape[0] == 0:
             # There is no proposal in the whole batch
             det_bboxes = [rois.new_zeros(0, 5)] * num_imgs
-            det_labels = [rois.new_zeros(0, )] * num_imgs
+            det_labels = [rois.new_zeros((0, ), dtype=torch.long)] * num_imgs
             det_bbox_results = [[
                 np.zeros((0, 5), dtype=np.float32)
                 for _ in range(self.bbox_head[-1].num_classes)
